move wsid to types
